requirements.txt
micloud
python-miio
requests
